Removed low pass filtering to improve timing

+++ report.h
@@ -64,7 +64,8 @@
 #define MESSAGE_DISABLED 5
 
 // Variable for holding voltage value after ADC
-extern float voltage_result[VOLTAGE_SENSOR_COUNT];
+//extern float voltage_result[VOLTAGE_SENSOR_COUNT];
+extern uint16_t voltage_result[VOLTAGE_SENSOR_COUNT];
 extern uint8_t voltage_result_index;
 
 // Prints system status messages.
